requirements.txt
streamlit
reportlab
pillow
cairosvg 